Refactored and optimized Accordion
