+++ src/components/accordion/styles.ts
@@ -1,0 +1,16 @@
+import {StyleSheet} from "react-native";
+
+export const componentStyles = StyleSheet.create({
+    title: {
+        flexDirection: "row",
+        alignItems: "center",
+    },
+    titleText: {
+        flex: 11,
+    },
+    titleBtn: {
+        flex: 1,
+        justifyContent: "center",
+        alignItems: "center",
+    },
+});
