+++ src/components/accordion/types.ts
@@ -1,13 +1,15 @@
 import type {Falsy, RecursiveArray, RegisteredStyle, ViewStyle} from "react-native";
 import type React from "react";
 
-export type AccordianProps = {
-    style?: ViewStyle | RegisteredStyle<ViewStyle> | RecursiveArray<ViewStyle | Falsy | RegisteredStyle<ViewStyle>>;
-    titleStyle?: ViewStyle | RegisteredStyle<ViewStyle> | RecursiveArray<ViewStyle | Falsy | RegisteredStyle<ViewStyle>> | readonly (ViewStyle | Falsy | RegisteredStyle<ViewStyle>)[];
-    title: React.ReactChild | React.ReactFragment;
+export type AccordionProps = {
+    styles: {
+        style?: ViewStyle | RegisteredStyle<ViewStyle> | RecursiveArray<ViewStyle | Falsy | RegisteredStyle<ViewStyle>>;
+        titleStyle?: ViewStyle | RegisteredStyle<ViewStyle> | RecursiveArray<ViewStyle | Falsy | RegisteredStyle<ViewStyle>> | readonly (ViewStyle | Falsy | RegisteredStyle<ViewStyle>)[];
+        bodyStyle?: ViewStyle | RegisteredStyle<ViewStyle> | RecursiveArray<ViewStyle | Falsy | RegisteredStyle<ViewStyle>> | readonly (ViewStyle | Falsy | RegisteredStyle<ViewStyle>)[];
+    };
+    title: React.ReactNode;
     showIcon?: boolean;
-    bodyStyle?: ViewStyle | RegisteredStyle<ViewStyle> | RecursiveArray<ViewStyle | Falsy | RegisteredStyle<ViewStyle>> | readonly (ViewStyle | Falsy | RegisteredStyle<ViewStyle>)[];
-    children: React.ReactChild | React.ReactFragment;
+    children: React.ReactNode;
     open?: boolean;
     setOpen?: () => null | void
 }
